+++ src/DI/ClassAliasMap.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types = 1);
+
+return [
+	\Kdyby\DoctrineCache\NotImplementedException::class 	=> \Kdyby\DoctrineCache\Exception\NotImplementedException::class,
+	\Kdyby\DoctrineCache\Exception::class 					=> \Kdyby\DoctrineCache\Exception\Exception::class,
+];
